# Conflicts:
#	ChangeLog
#	RecordTimer.py
#	configure.ac
#	data/menu.xml
#	data/setup.xml
#	data/vfd255/skin_display.xml
#	data/vfd255/skin_display_picon.xml
#	lib/base/eerror.cpp
#	lib/dvb/dvb.cpp
#	lib/dvb/dvb.h
#	lib/dvb/epgcache.cpp
#	lib/dvb/frontend.cpp
#	lib/gdi/picload.cpp
#	lib/gui/elistboxcontent.cpp
#	lib/python/Components/AVSwitch.py
#	lib/python/Components/About.py
#	lib/python/Components/ActionMap.py
#	lib/python/Components/ConfigList.py
#	lib/python/Components/Converter/MovieInfo.py
#	lib/python/Components/DiskInfo.py
#	lib/python/Components/EpgList.py
#	lib/python/Components/HdmiCec.py
#	lib/python/Components/InputDevice.py
#	lib/python/Components/Network.py
#	lib/python/Components/ParentalControl.py
#	lib/python/Components/Renderer/RollerCharLCD.py
#	lib/python/Components/ServiceScan.py
#	lib/python/Components/UsageConfig.py
#	lib/python/Components/config.py
#	lib/python/Plugins/Extensions/MediaPlayer/plugin.py
#	lib/python/Plugins/SystemPlugins/WirelessLan/Wlan.py
#	lib/python/Screens/About.py
#	lib/python/Screens/AudioSelection.py
#	lib/python/Screens/InfoBarGenerics.py
#	lib/python/Screens/InstallWizard.py
#	lib/python/Screens/MovieSelection.py
#	lib/python/Screens/NetworkSetup.py
#	lib/python/Screens/Rc.py
#	lib/python/Screens/ScanSetup.py
#	lib/python/Screens/ServiceInfo.py
#	lib/python/Screens/Setup.py
#	lib/python/Screens/SoftwareUpdate.py
#	lib/python/Screens/UserInterfacePositioner.py
#	lib/python/Screens/VideoMode.py
#	lib/python/Screens/Wizard.py
#	lib/python/Tools/Directories.py
#	lib/service/servicedvb.cpp
#	mytest.py
#	skin.py
#	tools/enigma2.sh.in

+++ lib/dvb/epgcache.cpp
@@ -1217,9 +1217,18 @@ void eEPGCache::load()
 	}
 	else
 	{
-		unlink(EPGDATX);
-		renameResult = rename(EPGDAT, EPGDATX);
-		if (renameResult) eDebug("[EPGC] failed to rename %s to %s: %m", EPGDAT, EPGDATX);
+		// only write to epg.dat if this is enabled in enigma2 config
+		if (!eConfigManager::getConfigBoolValue("config.epg.cachesaveenabled", false))
+		{
+			renameResult = -1;
+			eDebug("[eEPGCache::load] writing to %s is disabled in enigma2 config", EPGDAT);
+		}
+		else
+		{
+			unlink(EPGDATX);
+			renameResult = rename(EPGDAT, EPGDATX);
+			if (renameResult) eDebug("[EPGC] failed to rename %s to %s: %m", EPGDAT, EPGDATX);
+		}
 	}
 	{
 		int size=0;
@@ -1329,6 +1338,12 @@ void eEPGCache::save()
 	eDebug("[EPGC] save()");
 #endif
 	const char* EPGDAT = m_filename.c_str();
+	// only write to epg.dat if this is enabled in enigma2 config
+	if (!eConfigManager::getConfigBoolValue("config.epg.cachesaveenabled", false))
+	{
+		eDebug("[eEPGCache::save] writing to epg.dat is disabled in enigma2 config");
+		return;
+	}
 	if (eventData::isCacheCorrupt)
 		return;
 	// only save epg.dat if it is not empty
